EquationsOfStateOfSolids

Package Features

Calculate energy, pressure, and bulk modulus of an EOS Parameters on a (an) volume (array of volumes).

Fit an EquationOfState on a series of volumes using least-squares fitting method.

Fit an EquationOfState on a series of volumes linearly.

Find the corresponding volume of an EOS Parameters given an (a) energy, pressure, and bulk modulus.

Handle unit conversion automatically in the above features, take any unit.

See the Index for the complete list of documented functions and types.

Manual Outline

Installation

Installing Julia

on macOS

on Linux

Installing the package

Reinstall

How to make your data portable?

How to use EquationsOfStateOfSolids in Python?

Plotting

FAQ

How to make a Vector from a Parameters?

linfit does not work with BigFloat?

Collections

Usage

Construct a Parameters instance

Evaluate energy

Evaluate pressure

Evaluate bulk modulus

Public interfaces

Nonlinear fitting

Usage

Public interfaces

References

EquationsOfStateOfSolids.BirchMurnaghan

EquationsOfStateOfSolids.BirchMurnaghan

EquationsOfStateOfSolids.BirchMurnaghan2nd

EquationsOfStateOfSolids.BirchMurnaghan3rd

EquationsOfStateOfSolids.BirchMurnaghan4th

EquationsOfStateOfSolids.EnergyEquation

EquationsOfStateOfSolids.EnergyEquation

EquationsOfStateOfSolids.Murnaghan

EquationsOfStateOfSolids.PoirierTarantola

EquationsOfStateOfSolids.PoirierTarantola

EquationsOfStateOfSolids.PoirierTarantola2nd

EquationsOfStateOfSolids.PoirierTarantola3rd

EquationsOfStateOfSolids.PoirierTarantola4th

EquationsOfStateOfSolids.PressureEquation

EquationsOfStateOfSolids.PressureEquation

EquationsOfStateOfSolids.Vinet

EquationsOfStateOfSolids.Fitting.linfit

EquationsOfStateOfSolids.Fitting.nonlinfit

EquationsOfStateOfSolids.getparam

EquationsOfStateOfSolids.getparam

EquationsOfStateOfSolids.orderof

EquationsOfStateOfSolids.orderof

EquationsOfStateOfSolids.BirchMurnaghan — Method

BirchMurnaghan(args...)

Construct a BirchMurnaghan based on the length of arguments, where e0 must be provided.

See also: BirchMurnaghan2nd, BirchMurnaghan3rd, BirchMurnaghan4th

EquationsOfStateOfSolids.EnergyEquation — Method

(eos::EnergyEquation{<:PoirierTarantola3rd})(v)

Evaluate a Poirier–Tarantola energy EOS on volume v.

\[E(V) = E_0 + \frac {B_0 V_0}{2}\left[ \ln \left( \frac{V_0}{V}\right) \right]^{2} + \frac{B_0 V_0}{6} \left[ \ln \left(\frac{V_0}V{} \right) \right]^3 \left(B_0^\prime - 2 \right)\]

EquationsOfStateOfSolids.PoirierTarantola — Method

PoirierTarantola(args...)

Construct a PoirierTarantola based on the length of arguments, where e0 must be provided.

See also: PoirierTarantola2nd, PoirierTarantola3rd, PoirierTarantola4th

EquationsOfStateOfSolids.PressureEquation — Method

(eos::PressureEquation{<:PoirierTarantola3rd})(v)

Evaluate a Poirier–Tarantola pressure EOS on volume v.

\[P(V) = B_0 \frac{V_0}{V} \left[\ln \left( \frac{V_0}{V} \right) + \frac{\left( B_0^\prime -2 \right) }{2} \left[ \ln \left( \frac{V_0}{V} \right) \right]^2\right]\]

EquationsOfStateOfSolids.getparam — Method

getparam(eos::EquationOfStateOfSolids)

Get the Parameters from an EquationOfStateOfSolids.

EquationsOfStateOfSolids.orderof — Method

orderof(x::FiniteStrainParameters)

Return the order of a FiniteStrainParameters.

Examples

julia> orderof(BirchMurnaghan3rd(40, 0.5, 4, 0)) == 3
true