Finite strains

This module contains some methods to calculate several finite strains. The following formulae are from the Gibbs2 paper Table 3.

Eulerian strain:

\[f = \frac{ 1 }{ 2 } \bigg( \Big \frac{ V_0 }{ V } \Big^{\frac{ 2 }{ 3 }} - 1 \bigg)\]

Lagrangian strain:

\[f = \frac{ 1 }{ 2 } \bigg( \Big( \frac{ V }{ V_0 } \Big^{\frac{ 2 }{ 3 }} - 1 \bigg)\]

Natural (Hencky) strain:

\[f = \frac{ 1 }{ 3 } \ln \Big \frac{ V }{ V_0 } \Big\]

Infinitesimal strain:

\[f = 1 - \Big \frac{ V_0 }{ V } \Big^{\frac{ 1 }{ 3 }}\]

EquationsOfStateOfSolids.FiniteStrains.VolumeTo — Type

EulerianStrainFromVolume(v0)
LagrangianStrainFromVolume(v0)
NaturalStrainFromVolume(v0)
InfinitesimalStrainFromVolume(v0)

Calculate the finite strain of v based on the reference volume v0.

Examples

julia> f = EulerianStrainFromVolume(10);

julia> f(9)
0.036382991447572066

julia> f = EulerianStrainFromVolume(100u"nm^3");

julia> f(90u"nm^3")
0.036382991447572066

julia> g = inv(f);

julia> g ∘ f == f ∘ g == identity
true

EquationsOfStateOfSolids.FiniteStrains.VolumeFrom — Type

VolumeFromEulerianStrain(v0)
VolumeFromLagrangianStrain(v0)
VolumeFromNaturalStrain(v0)
VolumeFromInfinitesimalStrain(v0)

Calculate the original volume v from the finite strain f based on the reference volume v0.

Examples

julia> g = VolumeFromEulerianStrain(10);

julia> g(0.036382991447572066)
9.000000000000002

julia> g = VolumeFromEulerianStrain(100u"nm^3");

julia> g(0.036382991447572066)
90.00000000000001 nm³

julia> f = inv(g);

julia> f ∘ g == g ∘ f == identity
true